move scripts/cloud to src and scripts/utils to src/classifier
